RAG-embedding/requirements.txt
sentence-transformers
mysql-connector-python
mysql
tidb-vector[client]
pymysql
pillow
datasets